+++ LongPipesOrderForm/src/main/PipeT4.java
@@ -9,7 +9,7 @@ package main;
 public class PipeT4 extends PipeT3 {
 
     //costPerInch not needed, T3 and 4 share the same cost per inch.
-    public PipeT4(int grade, boolean chemicalResistance, double length, double outerDiameter, byte qty) {
+    public PipeT4(int grade, boolean chemicalResistance, double length, double outerDiameter, int qty) {
         super(grade, chemicalResistance, length, outerDiameter, qty);
         calculateCostT4();
     }
