+++ LongPipesOrderForm/src/main/PipeT5.java
@@ -8,7 +8,7 @@ package main;
 
 public class PipeT5 extends PipeT4 {
 
-    public PipeT5(int grade, boolean chemicalResistance, double length, double outerDiameter, byte qty) {
+    public PipeT5(int grade, boolean chemicalResistance, double length, double outerDiameter, int qty) {
         super(grade, chemicalResistance, length, outerDiameter, qty);
         costPerInch = new double[]{0.75, 0.8, 0.95};//set cost per inch
         calculateCostT5();
